[Feature #161420690] Create test for getting a single sale entry
