+++ app/test/v1/test_views.py
@@ -482,15 +482,7 @@ class TestsForApi(unittest.TestCase):
             'x-access-token': self.attendant_token['token']})
         self.assertEqual(response.status_code, 403)
 
-    def test_admin_get_single_sale(self):
-        response = self.test_client.get('/api/v1/sales/1', headers={
-            'x-access-token': self.admin_token['token']})
-        self.assertEqual(response.status_code, 200)
-
-    def test_attendant_get_single_sale(self):
-        response = self.test_client.get('/api/v1/sales/1')
-        self.assertEqual(response.status_code, 401)
-
+    
     def test_get_sales(self):
         response = self.test_client.get("/api/v1/sales",
                                         headers={
